+++ packages/jaspr_content/lib/src/data_loader/memory_data_loader.dart
@@ -7,11 +7,11 @@ import 'data_loader.dart';
 class MemoryDataLoader implements DataLoader {
   const MemoryDataLoader(this.data);
 
-  final Map<String, dynamic> data;
+  final Map<String, Object?> data;
 
   @override
   Future<void> loadData(Page page) async {
-    var pageData = page.data['page'] ?? {};
+    var pageData = page.data.page;
     page.apply(data: data);
     page.apply(data: {'page': pageData});
   }