+++ packages/jaspr_content/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added support for `sitemap: false` in the page data to exclude pages from the sitemap.
 - Ignores files and folders in the content directory starting with `.` (in addition to `_`).
 - Require Dart 3.6 or later to match the other Jaspr packages.
+- Renamed the `GithubButton` component to `GitHubButton`.
+- Fixed `GitHubButton` not linking to the correct repository.
 
 ## 0.2.0