+++ packages/jaspr_content/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added support for `sitemap: false` in the page data to exclude pages from the sitemap.
 - Ignores files and folders in the content directory starting with `.` (in addition to `_`).
 - Require Dart 3.6 or later to match the other Jaspr packages.
+- Renamed the `GithubButton` component to `GitHubButton`.
+- Fixed `GitHubButton` not linking to the correct repository.
 - Add a `namespacedPageData` extension to `Page` to access `page.data['page']`.
 - Switch the type of data in `Page` from `Map<String, dynamic` to `Map<String, Object?>`.